Add layout meta data
This prevents from MSKLC adding default values or even failing.

Do not add locale meta data, this can have the side effect that
installing the keyboard layout on a system without that
system locale present does add that system locale.
Without locale meta data MSKLC automatically sets the locale that
is active when compiling the layout and creating the installer.